refactor: separeted codebase
